CA-404597: rrd/lib_test - Verify that RRD handles non-rate data sources correctly
Other unit tests only verify the interoperability of the RRDs - dumping them to
JSON/XML and reading back in, verifying that the same data was decoded. We're
now seeing a problem where Gauge data sources, which should be absolute values
provided by the plugin, fluctuate wildly when processed by the RRD library.

Ensure we have an easy way to test this for both Gauge and Absolute data
sources - these values should be passed as-is by the RRD library, without any
time-based transformations.

This test currently fails and will be passing with the fix commits.

Signed-off-by: Andrii Sultanov <andrii.sultanov@cloud.com>