CP-53161: Link traces even when `smapi` component is experimental
This links requests back into `xapi` from `smapi` to the xapi request
that called the SM scripts in the first place.

When `smapi` was experimental there would be no traceparent header
passed in the request from `smapi` to `xapi`. Now, when `smapi` is
experimental and the tracing is enabled for xapi, the traceparent passed
to `smapi` is passed back to `xapi` for each request.

Signed-off-by: Gabriel Buica <danutgabriel.buica@cloud.com>