CA-404611: SXM: check power-state just before metadata export
A VM's power-state may change during a cross-pool storage migration, as
a VM that was originally running may shut down while the disks are being
copied (or vice versa). This case is handled, because copying disks can
take a very long time.

VM metadata is exported and imported into the remote pool after copying
has completed, and the VM power-state is used to decide whether to do
CPUID checks. Unfortunately, the power-state is currently taken from the
DB right at the start if the function, which does not take into account
that is may change. Straightforward fix.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>