CA-405404: Fix path to dracut
On all platforms tested the canonical location for dracut is `/usr/bin/dracut`.
Older platforms symlink to it from `/usr/sbin/dracut`, whilst newer ones do not,
thus on these platforms xapi is unable to find dracut and fails to start.

Adjust the path to the canonical location to ensure it operates correctly on all
platforms.

Also adjust a reference in network_utils.ml (whilst newer platforms also seem to
have the symlink from `/sbin/dracut`, it would seem to make sense to point to
the canonical location regardless).

Signed-off-by: Alex Brett <alex.brett@cloud.com>