CA-405593: Normalise API-installed host certificates
When installing host certificates, the parser used accepts strings with
extraneous characters surrounding the PEM-encoded data. The ad-hoc parser used
to reject such data because it was stricter.

Because the PEM-encoded objects are copied as-is after validating them, the
ad-hoc parser failed to read the file correctly on xapi restarts.

This change fixes the issue by encoding the key and certificates in PEM format
from parsed datastructures instead of using user-provided inputs directly.

    Parse, don't validate

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>