CA-405754: Update xapi-storage-script state.db
xapi-storage-script maintains a state.db which is loaded on startup
every time. This state.db contains currently attached SRs, and this
information can be used later on by xapi to decide whether to plug a PBD
in.

This information is correctly updated when adding an attached SR in, but
not when removing an SR. This will cause problems when an SR is
detached, followed by a toolstack restart. As now xapi-storage-script
still thinks that SR is attached, but in fact it is not.

Signed-off-by: Vincent Liu <shuntian.liu2@cloud.com>