[maintenance] blkgetsize64: avoid warning about uninit value
Analyzer can't figure out that 'rc' is set on 'fd < 0',
and we never try to use 'size_in_bytes'.

Appears to be a limitation in the analyzer, it doesn't realize that `fd < 0` implies that `rc` is set to `NOT_IMPLEMENTED`,
which raises an exception, so `caml_copy_int64` is never reached.

But the logic there is quite hard to follow, so better be explicit, and ensure `size_in_bytes` is always initialized.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>