Design proposal to support import/export of Qcow2 VDI
At XCP-ng, we are working on overcoming the 2TiB limitation
for VM disks while preserving essential features.
However, we currently cannot export a VDI to a Qcow2 file,
nor import one. The purpose of this design proposal is to outline
a solution for implementing VDI import/export in Qcow2 format.

Signed-off-by: Guillaume <guillaume.thouvenin@vates.tech>