Use Cmdliner for gen_api_main.exe
This removes the usage of the stdlib "Arg" module for parsing arguments
to the OCaml backend targets.

Using Cmdliner is more flexible and gives more opportunity for future
documentation or modification to build targets.

We remove global state and the ability to specify an output file: dune's
"with-stdout-to" action (and the piping capabilities of every shell)
subsumes this functionality.

Signed-off-by: Colin James <colin.barr@cloud.com>