CA-407328: Change vm parameter names of SXM calls
The vm parameter passed to SMAPIv2 calls is needed for the storage
backend to know which domain slice it should operate on. For SXM, since
there is no specific VM, we use the convention "MIR<vm-prefix>", where
the <vm-prefix> is the first five characters of the hash of the vm uuid.
This is not sufficient for VMs with multiple VDIs attached to it as they
would result in the same vm parameter.

So use a new convention "MIR<vm-prefix><vdi-prefix>" so we distinguish
VDIs during migration. These vm parameters cannot be too long though as
otherwise the storage backend will not accept them.

Signed-off-by: Vincent Liu <shuntian.liu2@cloud.com>