CA-407322 - libs/rrd: Keep lastupdate XML field as int, XenCenter relies on it
Instead of waiting for a complicated fix on XC side, switch back to outputting
integers when dumping XMLs.

Implementations that already switched to float should be good, since converting
int to float is fine, but not the other way around.

XML is an outdated format, JSON should be preferrably used, and it will keep
using floats with larger precision.

Fixes: a897b53 ("CA-391651: Use per-datasource last_updated timestamp during updating and archiving")

Signed-off-by: Andrii Sultanov <andrii.sultanov@cloud.com>