CP-54065, xenopsd: use domain_claim_pages on a single node, if possible
Xen currently supports to modes to claim memory for a domain: without any node
in particular, or claim memory in a single NUMA node.

When planning a domain, return the nodes that will host the domain, and how
much memory. In the case where the domain fits in a single NUMA node, claim
pages on that node, otherwise fall back to previous behaviour.

The memory claims need to happen while the memory measurements hold valid, that
is while no VMs are started, otherwise ENOMEM might be returned.

Because the current mode is a best-effort, log when the claim does not work.

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>