CI: fix compile_commands.json caching
When the cache for compile_commands.json got reused we didn't install
the OCaml compiler, which caused failures in the static analyzer,
because it couldn't locate the compiler runtime headers.

Do not skip setup-xapi-environment when the cache is reused.

This could be optimized to skip installing dune packages
(install just the compiler), but for now just install everything
to avoid the error.

Now when the cache is used we only skip installing/running
dune-compiledb.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>