CP-54827: sexpr: avoid escaping "
The non-standard S-expression serializer wraps all strings in `'`.
The lexer also ignores every non-escaped character after a `'`, until it sees the next
`'` or escape char.

So it should be safe to avoid escaping `"`.

Unescaping is unchanged: any character can be escaped with `\`,
and it returns it unchanged after removing the escape char, so this preserves backwards
compatibility when loading an old database.

Escaping shows up in performance profiles, because `trusted_on_first_use` field
in the pool contains a JSON, which contains \" characters that needed escaping.
With this change it won't anymore.

`ministat` confirms that there is an improvement:
```
sexpr_of_json_string:
    N           Min           Max        Median           Avg        Stddev
x 806     1438.7128     63663.127     1490.3661     1594.5095     2192.0548
+ 850     911.23529     48528.173     967.25054     1037.7674     1632.6855
Difference at 95.0% confidence
        -556.742 +/- 185.531
        -34.9162% +/- 9.28192%
        (Student's t, pooled s = 1925.34)

str_of_sexp_json:
    N           Min           Max        Median           Avg        Stddev
x 792     1622.9135     49591.388     1719.5377     1786.3412     1702.6472
+ 893     605.37329     3354.8035     626.51812     636.34457     107.24734
Difference at 95.0% confidence
        -1150 +/- 111.92
        -64.3772% +/- 2.26583%
        (Student's t, pooled s = 1169.88)
```

Signed-off-by: Edwin Török <edwin.torok@cloud.com>