CP-53951: Drop SSL and Lwt dependency from XAPI
XAPI dependended on vhd_lib, which depends on ssl and lwt.
XAPI is not yet ready for Lwt (and mixing Lwt with blocking code is not a good idea).

The library was only used to gain access to 1 variant used for parameter passing inside XAPI.
Replace this with a polymorphic variant, and drop the dependency.
Also allows to drop a match case that was dead code (proto was only ever set to NBD).

Fixes: 044dc15ee4 ("CP-45016: Add support for specifying nbd export in sparse_dd")

Signed-off-by: Edwin Török <edwin.torok@cloud.com>