CA-407177: Fix swtpm's use of SHA1 on XS9
The default crypto policy in XS9 disables use of SHA1. However, swtpm
needs to use it since it advertises SHA1 support to guests. On XS9,
swtpm will ship with a custom openssl configuration file for this
purpose so set the appropriate environment variable to use it if the
file exists.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>