CP-54034: Expose `expected_votes` in Cluster object
The `expected_votes` field in corosync represents the number of hosts
that is expected by the cluster stack. In the context of corosync, this
is the same as the number of hosts as in the corosync.conf file*. This
is a useful field to expose to the user so that they can see how many
nodes actually are expected.

We also have `Cluster_host` object, which represents xapi's view of what
nodes should be in the cluster, but that might not be identical to
corosync's view, especially when a host is disabled, but is still left
in the list of Cluster_host objects.

Although one could argue that we could infer this `expected_votes` field
from the number of enabled Cluster_hosts, it might still be useful to
get this information directly from corosync.

*: there are ways in corosync to make one host cast multiple votes, but
that feature is not used.

Signed-off-by: Vincent Liu <shuntian.liu2@cloud.com>