CP-54026: option to control VM-internal shutdown behaviour under HA
Adds a new datamodel field in the pool class:

    bool pool.ha_reboot_vm_on_internal_shutdown

The field is read/write, with the setter restricted to the pool-operator
role. The default of this new field is `true` to reflect the current
behaviour.

This field controls what happens when HA is enabled and a protected VM
is shut down internally (e.g. by pressing the shutdown button in
Windows):

- `true`: the VM is automatically restarted.
- `false`: the VM is not restarted and left halted – consistent with
  the behaviour of shutting the VM down through the API.

CLI:

    xe pool-param-set uuid=... ha-reboot-vm-on-internal-shutdown=false

Whether an HA-protected VM is automatically (re)started depends on the
field `VM.ha_always_run`, which is managed by xapi. This field is set to
`true` when a protected VM is started, and to `false` when it is shut
down through the API, which prevents the HA monitor thread from
restarting it again. Setting the new pool-field to `false` does the
same thing is such a VM is shut down from inside when handling the event
in the xenopsd-events thread.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>