CP-44103: Ordering network devices
This part sorts host network devices in xcp-networkd. Previously, the
ordering was handled by the interface-rename functionality. This will
now be replaced by an equivalent function in xcp-networkd, but without
renaming the network devices. The renmaing performed by the
interface-rename was used to record the sorting result as the name of
the NICs like "eth<N>". Now the sorting result will be recorded as
internal files.

Signed-off-by: Ming Lu <ming.lu@cloud.com>