CP-53555: Split unplug atomic into deactivate/detach
This consists of two parts, splitting the unplug function into
functionally equivalent deactivate and detach functions, and splitting
the VBD_unplug atomic itself's behaviour into two new atomics:
VBD_deactivate and VBD_detach

If the xenopsd_vbd_plug_unplug_legacy flag is true, the only difference
will be that VBD_unplug calls deactivate and detach sequentially instead
of unplug

If xenopsd_vbd_plug_unplug_legacy is false, the VBD_deactivate and
VBD_detach atomics will be used in place of VBD_unplug in all places
that it is used. This should still be functionally equivalent.

The purpose of this change is to allow optimisations in this area as
there are some situations where they do not need to be called at the
same time. For example we could skip detaching on reboot and only
deactivate and activate again reducing reboot time.

Signed-off-by: Steven Woods <steven.woods@cloud.com>