xapi_vm_migrate: Avoid duplicate, overly-strict CBT check on VDIs
There is already a call to `assert_can_migrate_vdis` present in
`assert_can_migrate`, which checks that none of the VDIs that *are going to be
moved* have CBT enabled. There is no need to additionally check that none of the
VDIs *in general* have CBT enabled.

Some clients, like XenOrchestra, will turn off CBT on VDIs and retry migration
after getting the `VDI_CBT_ENABLED` error on live migration. Dropping this
overly strict check allows not stripping CBT when VDI will not be moved (when
it's on a shared SR).
In addition, during rolling pool upgrades, disabling CBT is not allowed, hence
the live migration operation wouldn't be able to continue. Avoiding the strict
check fixes that as well.

Signed-off-by: Andrii Sultanov <andriy.sultanov@vates.tech>