Move the copying function to storage_smapiv1_migrate
Although copying is not exclusively used by SMAPIv1, e.g.
DATA.copy maps to Storage_migrate.copy which maps to
Storage_smapiv1_migrate.copy. And this is used when a VDI does not need
to be migrated live (attached read-only). It's a bit easier in terms of
dependency to leave them in storage_smapiv1_migrate.ml

And some related helper functions such as `progress_callback`.

Signed-off-by: Vincent Liu <shuntian.liu2@cloud.com>