Split Storage_migrate.start
The original Storage_migrate.start function is quite fat and does lots
of things. For better readability and maintainability, split it into
multiple functions that represent different stages of the SXM:
- prepare: makes a remote call to the dest host and asks it to prepare
  for the VDIs for receiving data
- mirror_pass_fds: passes the fd to tapdisk to establish connection
  between two tapdisk processes (sending & receiving)
- mirror_snapshot: takes a snapshot of the VDI to be mirrored, because
  the tapdisk mirroring only mirrors data written after the mirror is
  initiated
- mirror_copy: copy the snapshot from the source to destination

As these operations are all specific to SMAPIv1, move them to
`storage_smapiv1_migrate` as well.

Also restructure the way clean up is done. As there are multiple points
at which the migration might fail, previously a `on_fail` list ref is
constructed and populated as we go. Now that we have explicitly defined
the different errors during the migration, we could handle these errors
accordingly.

The way to reason about cleanups is: we have three possible clean up
actions:

- receive_cancel: which will undo the preparation work done on the dest
  host. This is added once preparation is done;
- destroy snapshot: which will destroy the snapshot. This is called
  after the snapshot is created;
- stop: which will disable mirroring, delete snapshot and also do what
  receive_cancel does, which is basically clean up everything. And this
  will be called after the mirror has been started.

Note many of these functions are best-effort, for example, the stop
function won't delete a snapshot if it does not exist, so it's generally
safe to call them.

We put these clean up functions into the appropriate places, see the
actual arrangement of these functions in the code, they should be
self-explanatory.

Signed-off-by: Vincent Liu <shuntian.liu2@cloud.com>