CA-401023: Remove smapi observer config if smapi is set as experimental
If smapi observer was previously enabled and then added to
observer_experimental_components, when the toolstack is restarted the
smapi observer will no longer exist. The observer will therefore not be
"destroyed" by normal means so we must remove the config manually.

This is important for xapi-storage-script as it cannot read the DB and
so relies on this config file to know whether to trace SMAPIv3 or not.

Signed-off-by: Steven Woods <steven.woods@cloud.com>