CA-409488: prevent Xenctrl exceptions from escaping on VM boot/shutdown races
Note that this hostload metric is completely wrong, and we should
probably drop it.
But for now prevent it from introducing gaps in the entire host CPU
graph: if an exception escapes here, then nothing else gets reported,
because the host CPU, VM CPU and hostload metrics all get returned as a single
list. An exception escaping from any prevents all the others from
working too.

Fixes: 9aa7dfcca ("CP-43574: Add host load data source")
Signed-off-by: Edwin Török <edwin.torok@cloud.com>