CA-409489: prevent running out of pages with lots of domains
When this plugin was part of xcp-rrdd the shared memory protocol wasn't
used for transfering data (it was transferred by direct function call).
The shared memory protocol has a limitation that you need to declare a
maximum size from the beginning, or carefully grow the file itself
dynamically.

We support max 1024 VMs/host, and if each one needs 5 pages,
that'd be 20Mib/host, so always use that as the minimum number of pages.

Fixes: b3ea09222 ("IH-615: Move CPU-related data-source collection into a separate RRDD plugin")
Signed-off-by: Edwin Török <edwin.torok@cloud.com>