CA-408230: Enable destroy op for HA statefile VDI after HA is disabled
The destroy operation is missing in the allowed operations for HA
statefile VDI after HA is disable. It is because when
update_allowed_operations for the VDI, it is found that
ha_disable_in_progress is true.

Xapi_ha.disable_internal will also be called when HA is not enabled
successfully, in that case, ha_enable_in_progress is true.

This fix removes `ha_enable or `ha_disable from pool's current
operations to allow update_allowed_operations to enable destory
operation for the VDI.

Signed-off-by: Gang Ji <gang.ji@cloud.com>