CP-52131/CP-53474: Reorder operations during pci_add
Reorder the operations in _pci_add:

1) So that Xen can verify calls to grant ioport and iomem permissions,
reorder the calls so that the device is assigned to the domain before
granting permissions for the resources. When Secure Boot is enabled, Xen
will enforce that ioport/iomem permissions can be granted to a domain
only when the corresponding device is assigned to that domain.

2) Add the device to QEMU after assigning to the domain. Rather than
accessing the PCI config space through dom0 sysfs which is blocked when
Secure Boot is enabled, QEMU in XS9 has been updated to use a hypercall
to access the PCI config space of a device assigned to a domain.
Therefore, add the device to QEMU after assigning it to the domain
rather than before so that the config space accesses it performs during
the QMP call succeed.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>