Add more states for SXM
Adding a few more states that will be tracked by the send_state and
receive_state. These will be used later on for SMAPIv3 outbound
migration.

We do need to give them a default value in case the deserialisation
of the on-disk copy does not have these values.

These are not used yet, so no functional change.

Signed-off-by: Vincent Liu <shuntian.liu2@cloud.com>