Remove receive_start(2) from storage_migrate
These functions will not be exposed in storage_migrate.ml any more.
`receive_start` is kept in the mux for backwards compatability reasons,
and will go straight to the SMAPIv1 implemenetation.

`receive_start2` calls will not be directed to the storage_mux any more,
instead it will be multiplexed by the new layer in storage_migrate.

Signed-off-by: Vincent Liu <shuntian.liu2@cloud.com>