CP-307922: Implement SMAPIv3 outbound migration
This is the main commit that implements the MIRROR interface in
storage_smapiv3_migrate. The exact detail of how SMAPIv3 mirror is done
is left in the SXM documentation, but core of it is to provide all the
necessary infrastructure to able to call the `Data.mirror` SMAPIv3 call
that will mirror a VDI to another.

Signed-off-by: Vincent Liu <shuntian.liu2@cloud.com>