CA-409628: Do not lose the original backtrace in log_backtrace
There are 2 log_backtrace functions in the Debug module, one of them prints the full backtrace,
the other one prints the backtrace just from the last time Backtrace.is_important got called,
i.e. it drops the *important* part of the backtrace.

Delete the buggy function and replace it with a function that always takes an exception,
so that we can look up any stashed backtrace.

There are good reasons why Backtrace.is_important wipes the current backtrace after it stashes it away (destroying the important part):
 * if it didn't, then every time Backtrace.is_important got called we'd get the same backtrace prefix appeneded multiple times
 * for cross-process/language backtraces to work you need to use Backtrace.get to retrieve the stashed backtrace when printing, and not use the OCaml one directly from Printexc.

 But that means you need to be disciplined in how you deal with backtraces:
 * catching and reraising the same exception with the same backtrace using Printexc.rawbacktrace is fine
 * the first statement in an exception handler needs to be either Backtrace.is_important, or Debug.log_backtrace (which calls is_important internally)

 Using Printexc.get_backtrace/Printexc.get_rawbacktrace for printing purposes has to be avoided if we ever called Backtrace.is_important.

 The updated unit test shows that we properly print line 8 as the source of the exception now.

 This fixes backtraces from xcp-rrdd.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>