CA-409628: remove unnecessary log_backtrace from API boundary
We should only log_backtrace if we are the final handler.

The exception is raised here, so the caller will have a chance to log it.
This was also inconsistent: some *_interface logged the backtrace, and others didn't.

In theory there is a chance that the caller is buggy and doesn't log the correct backtrace.
But if we simplify the places that call the Backtrace module, we'll have fewer chances of that going wrong.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>