CA-409628: add missing Backtrace.is_important
This shows how brittle the current Backtrace API is,
this was missing from a lot of places.

We have some better alternatives (`with_backtraces`,
or a `try_with` function) that'd guarantee that `important` is
always called in the right place, but that would be a more invasive change,
which will be done in a followup commit.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>