Add Nbd parameters into VDI.attach
Whilst it is not the default behaviour on XS 8 to attach a VDI through
NBD, SXM inbound into a SMAPIv1 SR needs to have nbd enabled for
mirroring purposes. As tapdisk will return usable nbd parameters to
xapi, they can be included in the return value of attach. Most current
users of this return value will keep using blktap2 kernel device and
this nbd information is only used during SXM.

Signed-off-by: Vincent Liu <shuntian.liu2@cloud.com>