Move attach/activate for SXM
The attach and activate of the VDI being live migrated is there so that
the SXM can keep working even if the VM on which the VDI is activated
shutsdown. This is possible on SMAPIv1 as tapdisk does not distinguish
between different domain paramters. But that is not the case for
SMAPIv3.

For now just avoid activating the VDI on dom0 since the VM is already
activated on the live_vm. This does mean that SXM will stop working if
the VM is shut down during storage migration. We will leave that case in
the future.

Signed-off-by: Vincent Liu <shuntian.liu2@cloud.com>