Bring back DATA.MIRROR.list to storage_mux
Previously this was deleted in commit 1fe6389 as it was not multiplexed,
but looks like we still need to keep it in storage_mux because sm-cli
needs to make rpc calls to storage_mux when trying to list all the
mirrors to make it work properly, due to the fact that the sr plugins
are stored in the address space of the xapi process.

There are other invocations in sm-cli such as `Storage_migrate.start`
which may have similar problems. But I have left them alone as I don't
any reasonable way of calling them from the cli.

Signed-off-by: Vincent Liu <shuntian.liu2@cloud.com>